refactor(reservation-application): 방 운영 정책 구조 정리
- RoomOperationReservationPolicy를 기본 구현체 없이 직접 정책 객체로 사용하도록 정리함.

- RoomPolicyReason을 contract 패키지로 올리고 별도 result 래퍼를 제거함.

- 방 운영 정책 테스트를 현재 정책 객체 이름과 패키지에 맞춰 수정함.
