+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/room/contract/RoomPolicyReason.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.reservation.application.room.contract.result;
+package com.seatliberator.seatliberator.reservation.application.room.contract;
 
 import com.seatliberator.seatliberator.reservation.application.shared.policy.PolicyDecision;
 import com.seatliberator.seatliberator.reservation.application.shared.policy.PolicyReason;
